+++ Library/OcMainLib/OpenCoreUefi.c
@@ -298,10 +298,8 @@ OcReinstallProtocols (
   )
 {
   CONST CHAR8   *AppleEventMode;
-  CONST CHAR8   *CustomDelaysMode;
   BOOLEAN       InstallAppleEvent;
   BOOLEAN       OverrideAppleEvent;
-  BOOLEAN       UseCustomDelays;
 
   if (OcAudioInstallProtocols (Config->Uefi.ProtocolOverrides.AppleAudio) == NULL) {
     DEBUG ((DEBUG_INFO, "OC: Disabling audio in favour of firmware implementation\n"));
@@ -349,7 +347,6 @@ OcReinstallProtocols (
 
   InstallAppleEvent   = TRUE;
   OverrideAppleEvent  = FALSE;
-  UseCustomDelays     = Config->Uefi.Input.KeySupport;
 
   AppleEventMode = OC_BLOB_GET (&Config->Uefi.AppleInput.AppleEvent);
 
@@ -362,22 +359,10 @@ OcReinstallProtocols (
     DEBUG ((DEBUG_WARN, "OC: Invalid AppleInput AppleEvent setting %a, using Auto\n", AppleEventMode));
   }
 
-  CustomDelaysMode = OC_BLOB_GET (&Config->Uefi.AppleInput.CustomDelays);
-
-  if (AsciiStrCmp (CustomDelaysMode, "Auto") == 0) {
-  } else if (AsciiStrCmp (CustomDelaysMode, "Enabled") == 0) {
-    UseCustomDelays = TRUE;
-  } else if (AsciiStrCmp (CustomDelaysMode, "Disabled") == 0) {
-    UseCustomDelays = FALSE;
-  } else {
-    DEBUG ((DEBUG_WARN, "OC: Invalid AppleInput CustomDelays setting %a, using Auto\n", CustomDelaysMode));
-  }
-
-
   if (OcAppleEventInstallProtocol (
     InstallAppleEvent,
     OverrideAppleEvent,
-    UseCustomDelays,
+    Config->Uefi.AppleInput.CustomDelays,
     Config->Uefi.AppleInput.KeyInitialDelay,
     Config->Uefi.AppleInput.KeySubsequentDelay,
     Config->Uefi.AppleInput.PointerSpeedDiv,