+++ Changelog.md
@@ -3,6 +3,11 @@ OpenCore Changelog
 #### v0.6.9
 - Fixed out-of-sync cursor movement rectangle when loading e.g. CrScreenshotDxe
 - Updated underlying EDK II package to edk2-stable202102
+- Applied consistent enforcement of required minimum Apple OEM Apple Event protocol version
+- Changed CustomDelays to less surprising boolean setting with failsafe of false
+- Changed key repeat failsafes and sample values to Apple OEM values
+- Changed PointerSpeedMul failsafe to Apple OEM value
+- Updated docs to include configuration of key repeat settings on AMI KeySupport vs. V1 or V2 KeySupport vs. systems not using KeySupport
 
 #### v0.6.8
 - Switched to VS2019 toolchain for Windows builds